+++ common/src/main/java/club/iananderson/pocketgps/items/properties/GpsItemProperties.java
@@ -1,6 +1,8 @@
 package club.iananderson.pocketgps.items.properties;
 
 import club.iananderson.pocketgps.PocketGps;
+import club.iananderson.pocketgps.items.components.GpsToggle;
+import club.iananderson.pocketgps.items.components.ItemEnergy;
 import club.iananderson.pocketgps.util.NBTUtil;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.item.ClampedItemPropertyFunction;
@@ -12,7 +14,7 @@ public class GpsItemProperties implements ClampedItemPropertyFunction {
   @Override
   public float unclampedCall(ItemStack itemStack, @Nullable ClientLevel clientLevel,
       @Nullable LivingEntity livingEntity, int i) {
-    return (!NBTUtil.getBoolean(itemStack, PocketGps.TOGGLE_GPS_TAG) || (PocketGps.gpsNeedPower()
-        && NBTUtil.getInt(itemStack, PocketGps.ENERGY_TAG) == 0)) ? 0F : 1F;
+    return (!NBTUtil.getBoolean(itemStack, GpsToggle.TOGGLE, GpsToggle.initState) || (PocketGps.gpsNeedPower()
+        && NBTUtil.getInt(itemStack, ItemEnergy.ENERGY) == 0)) ? 0F : 1F;
   }
 }