+++ common/src/main/java/club/iananderson/pocketgps/util/ItemUtil.java
@@ -1,20 +1,20 @@
 package club.iananderson.pocketgps.util;
 
-import club.iananderson.pocketgps.PocketGps;
+import club.iananderson.pocketgps.items.components.GpsToggle;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 
 public class ItemUtil {
   public static void initGpsState(ItemStack itemStack) {
-    NBTUtil.setBoolean(itemStack, PocketGps.TOGGLE_GPS_TAG, true);
+    NBTUtil.setBoolean(itemStack, GpsToggle.TOGGLE, GpsToggle.initState);
   }
 
   public static boolean isGpsOn(ItemStack stack) {
-    return NBTUtil.getBoolean(stack, PocketGps.TOGGLE_GPS_TAG);
+    return NBTUtil.getBoolean(stack, GpsToggle.TOGGLE, GpsToggle.initState);
   }
 
   public static void toggleGps(ItemStack stack, Player player) {
-    boolean current = NBTUtil.getBoolean(stack, PocketGps.TOGGLE_GPS_TAG);
-    NBTUtil.flipBoolean(stack, PocketGps.TOGGLE_GPS_TAG);
+    boolean current = NBTUtil.getBoolean(stack, GpsToggle.TOGGLE, GpsToggle.initState);
+    NBTUtil.flipBoolean(stack, GpsToggle.TOGGLE, GpsToggle.initState);
   }
 }