+++ fabric/src/main/java/club/iananderson/pocketgps/fabric/PocketGpsFabric.java
@@ -1,11 +1,12 @@
 package club.iananderson.pocketgps.fabric;
 
 import club.iananderson.pocketgps.PocketGps;
-import club.iananderson.pocketgps.client.PocketGpsClient;
+import club.iananderson.pocketgps.config.PocketGpsConfig;
 import club.iananderson.pocketgps.fabric.registry.FabricRegistration;
 import club.iananderson.pocketgps.impl.accessories.AccessoriesCompat;
+import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeConfigRegistry;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
+import net.neoforged.fml.config.ModConfig.Type;
 
 public final class PocketGpsFabric implements ModInitializer {
   //Todo ---- Find out why it wont receive power on servers only
@@ -14,6 +15,9 @@ public final class PocketGpsFabric implements ModInitializer {
   public void onInitialize() {
     PocketGps.init();
     FabricRegistration.register();
+
+    NeoForgeConfigRegistry.INSTANCE.register(PocketGps.MOD_ID, Type.COMMON, PocketGpsConfig.GENERAL_SPEC,
+                                             "pocketgps-common.toml");
 
     if (PocketGps.accessoriesLoaded() && !PocketGps.trinketsLoaded()) {
       PocketGps.LOG.info("Talking to Accessories");